Add tests for is_number
