+++ microservices/store/tests/unit/test_handler.py
@@ -1,5 +1,4 @@
 import json
-import os
 import pytest
 import sys
 
@@ -8,25 +7,7 @@ sys.path.append("..")
 from src import app
 
 
-@pytest.fixture()
-def apigw_event(request):
-    # Mock API Gateway event
-    current_dir = os.path.dirname(__file__)
-    event_path = os.path.join(current_dir, "..", "..", "events", request.param)
-
-    with open(event_path, "r") as event_file:
-        return json.load(event_file)
-
-
-def test_get_all_aws_services(dynamodb_table, table_name):
-    # Given a DynamoDB table
-    # When a GET / request is called
-    items = app.get_aws_services(table_name)
-    # Then the entire table is returned
-    assert len(items) > 0
-    assert items[0] == dynamodb_table
-
-
+# Tests related to the main handler function
 @pytest.mark.parametrize("apigw_event", ["dev.json", "prod.json"], indirect=True)
 def test_lambda_handler(apigw_event):
     # Given an API Gateway event
